+++ packages/core/tests/js/vite/default/output.ts
@@ -1,0 +1,5 @@
+import myPlugin from 'my-plugin';
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vite';
+
+export default defineConfig({ plugins: [sveltekit(), myPlugin()] });
