+++ packages/core/tests/js/vite/with-variable/input.ts
@@ -1,0 +1,8 @@
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vite';
+
+const config = defineConfig({
+	plugins: [sveltekit()]
+});
+
+export default config;
