+++ packages/core/tooling/js/vite.ts
@@ -1,5 +1,4 @@
 import { array, functions, imports, object, exports, type AstTypes, common } from './index.ts';
-import { parseScript } from '../parsers.ts';
 
 export function exportDefaultConfig(
 	ast: AstTypes.Program,
@@ -75,17 +74,9 @@ export function addInArrayOfObject(
 }
 
 export const addPluginToViteConfig = (
-	content: string,
-	fn: (
-		ast: AstTypes.Program,
-		o: {
-			configObject: AstTypes.ObjectExpression;
-			add: (options: { code: string; mode?: 'append' | 'prepend' }) => void;
-		}
-	) => void
-): string => {
-	const { ast, generateCode } = parseScript(content);
-
+	ast: AstTypes.Program,
+	fn: (o: { add: (options: { code: string; mode?: 'append' | 'prepend' }) => void }) => void
+): void => {
 	// Step 1: Get the config object, or fallback.
 	imports.addNamed(ast, { from: 'vite', imports: { defineConfig: 'defineConfig' } });
 	const configObject = exportDefaultConfig(ast, {
@@ -94,14 +85,12 @@ export const addPluginToViteConfig = (
 	});
 
 	// Step 2: Add the plugin to the plugins array
-	fn(ast, {
-		configObject,
-		add: (options) =>
+	fn({
+		add: (options) => {
 			addInArrayOfObject(configObject, {
 				array: 'plugins',
 				...options
-			})
+			});
+		}
 	});
-
-	return generateCode();
 };
